rename app -> store
