+++ package.json
@@ -23,21 +23,21 @@
     "date-fns": "^2.29.3",
     "feed": "^4.2.2",
     "flowbite": "^1.5.3",
-    "flowbite-react": "^0.2.0",
+    "flowbite-react": "^0.3.2",
     "globby": "^13.1.2",
     "gray-matter": "^4.0.3",
     "image-size": "^1.0.2",
     "lodash": "^4.17.21",
     "lqip-modern": "^1.2.0",
     "markdown-it": "^13.0.1",
-    "next": "^12.3.1",
-    "next-auth": "^4.14.0",
+    "next": "^13.0.0",
+    "next-auth": "^4.15.0",
     "next-contentlayer": "^0.2.8",
     "next-mdx-remote": "^4.1.0",
     "next-redux-wrapper": "^8.0.0",
     "next-remote-watch": "^2.0.0",
     "nextjs-progressbar": "^0.0.16",
-    "patch-package": "^6.4.7",
+    "patch-package": "^6.5.0",
     "process": "^0.11.10",
     "react": "^18.2.0",
     "react-async": "^10.0.1",
@@ -62,18 +62,19 @@
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/typography": "^0.5.7",
     "@types/jsdom": "^20.0.0",
-    "@types/node": "^18.11.2",
-    "@types/react": "^18.0.21",
-    "@typescript-eslint/eslint-plugin": "^5.40.1",
+    "@types/node": "^18.11.7",
+    "@types/react": "^18.0.23",
+    "@typescript-eslint/eslint-plugin": "^5.41.0",
     "autoprefixer": "^10.4.12",
-    "eslint": "^8.25.0",
-    "eslint-config-next": "^12.3.1",
+    "concurrently": "^7.5.0",
+    "eslint": "^8.26.0",
+    "eslint-config-next": "^13.0.0",
     "jsdom": "^20.0.1",
     "postcss": "^8.4.18",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
     "prisma": "^4.5.0",
-    "tailwindcss": "^3.1.8",
+    "tailwindcss": "^3.2.1",
     "typescript": "^4.8.4"
   }
 }
